remove useless functions and files
